Update README.md
We're currently attempting to retrieve data from files located within a SharePoint folder. We're utilizing requests_ntlm for authentication, which successfully returns a 200 response. However, we're encountering difficulties in extracting the filenames or data contained within files in the folder path from the response.content. We're utilizing BeautifulSoup() to parse the content and extract the necessary data from the files within the folder path. We'd appreciate any guidance or suggestions you could offer to improve the code for fetching data from the files within the SharePoint folder path